average added

+++ python/assignment_4/average.py
@@ -1,4 +1,4 @@
-numbers = list(map(int,input("Enter numbers using space:\n").split()))
+numbers = list(map(float,input("Enter numbers using space:\n").split()))
 length = len(numbers)
 average = sum(numbers)/length
-print("The Average of your numbers {} is {}".format(numbers,average))
+print("The Average of your numbers {} is {}".format(numbers, average))
